resize_root: Get size by taking end address of last partition
Previously, resize_root would calculate the size for the root partition
by summing the sizes of the partitions we were "eating." This works fine
when the partitions are contiguous. However, recently, I've been seeing
that cloudlab has partitions that are not contiguous, so that strategy
can leave a lot of disk capacity unclaimed. Taking the end address of
the last partition avoids this problem